Fix DEF_008: Integration Test Type/Format Assertion Mismatches - Fixed UICoreServiceIntegrationTest trader ID comparison (Int to Long) - Fixed StatePersistenceTest BigDecimal comparison using compareTo() - Both tests now pass - Updated defect document with root cause and solution - Related to Epic 6, Issue #25

+++ 03_Development/Application_OnPremises/core-service/src/integrationTest/kotlin/com/fmps/autotrader/core/integration/UICoreServiceIntegrationTest.kt
@@ -216,7 +216,7 @@ class UICoreServiceIntegrationTest {
         assertNotNull(apiResponse.data, "Response data should not be null")
         
         val trader = apiResponse.data!!
-        assertEquals(traderId, trader.id)
+        assertEquals(traderId.toLong(), trader.id)
         assertEquals(20, trader.leverage)
         
         println("✅ Trader configuration updated via REST API")
